backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.1
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
requests==2.31.0
aiohttp==3.9.3
sendgrid==6.11.0